+++ src/Luban.Job.Common/Source/TypeVisitors/TagNameVisitor.cs
@@ -118,7 +118,7 @@ namespace Luban.Job.Common.TypeVisitors
 
         public string Accept(TDateTime type)
         {
-            return "INT";
+            return "LONG";
         }
     }
 }
